api/requirements.txt
aiohttp==3.7.4.post0
aiohttp-middlewares==1.1.0
aiohttp-devtools==0.13.1
aiohttp-debugtoolbar==0.6.0
aiohttp-security==0.4.0
pylint==2.14.5
boto3==1.24.46
dnspython==1.16.0
Pillow==8.4.0
pymongo==3.12.0
PyYAML==5.4.1
pycodestyle==2.7.0
bcrypt==3.2.0
PyJWT==2.1.0
httpagentparser==1.9.1
